acpi/bus: Initialize arch specific platform devices
The RISC-V APLIC is a wired to MSI interrupt converter device
which is a platform device. However, this is also a type of
interrupt controller and hence it is enumerated using MADT
table. This is somewhat similar to how SMMU is initialized
using IORT on ACPI based systems.

Instead of adding individual platform device initialization,
add a weak function which can be defined by each platform. This
approach allows to add IOMMU device also in future without
modifying the common path.

Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>