RISC-V:irq.c: Initialize RISC-V INTC information
Information in the RISC-V INTC structure in MADT are referred by other
optional interrupt controllers like IMSIC, APLIC and PLIC. So, we need to
initialize the information about all of the RINTC structures much earlier
than those interrupt controller drivers get probed. The RINTC structures
should be saved in the same order as they appear in MADT including those
with disabled flag set. Hence, it is different from the existing
cpu_madt_rintc map which is based on cpuid as the index.

Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>