ACPI: scan.c: Define weak function to populate dependencies
RISC-V needs to define dependency between interrupt controller and the
devices which use GSI interrupts so that they are probed in the correct
order. This dependency already exists as part of GSI number mapping.
Hence, instead of depending on _DEP, the architecture specific function
can add this dependency based on GSI mapping. Add this weak architecture
function which can be defined by each architecture.

Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>