RISC-V: ACPI: Implement function to add auto dependencies
The APLIC in RISC-V which provides GSI interrupts is primarily wired
interrupt to MSI interrupt bridge. Since it needs MSI services, it can
not be probed early during boot and hence dependency needs to be added
between APLIC and the devices connected to it which use GSIs. Since the
dependency already exists as part of GSI number mapping, there is no
need for the firmware to add once again _DEP between each device and the
APLIC. Hence, implement the function to populate the dependency list
looking at the IRQ dependencies.

Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>