arm64/mm: pmd_mkinvalid() must handle swap pmds
__split_huge_pmd_locked() can be called for a present THP, devmap or
(non-present) migration entry. It calls pmdp_invalidate()
unconditionally on the pmdp and only determines if it is present or not
based on the returned old pmd.

But arm64's pmd_mkinvalid(), called by pmdp_invalidate(),
unconditionally sets the PMD_PRESENT_INVALID flag, which causes future
pmd_present() calls to return true - even for a swap pmd. Therefore any
lockless pgtable walker could see the migration entry pmd in this state
and start interpretting the fields (e.g. pmd_pfn()) as if it were
present, leading to BadThings (TM). GUP-fast appears to be one such
lockless pgtable walker.

While the obvious fix is for core-mm to avoid such calls for non-present
pmds (pmdp_invalidate() will also issue TLBI which is not necessary for
this case either), all other arches that implement pmd_mkinvalid() do it
in such a way that it is robust to being called with a non-present pmd.
So it is simpler and safer to make arm64 robust too.

This is a theoretical bug found during code review. I don't have any
test case to trigger it in practice.

Fixes: 53fa117bb33c ("arm64/mm: Enable THP migration")
Cc: <stable@vger.kernel.org> # 5.10.x
Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>
Acked-by: Will Deacon <will@kernel.org>
Reviewed-by: Zi Yan <ziy@nvidia.com>
Link: https://lore.kernel.org/r/20240430133138.732088-1-ryan.roberts@arm.com
[catalin.marinas@arm.com: removed the debug_vm_pgtable.c changes]
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>