Add SMMC driver for MSI only platforms with gas structure
The draft specification has opens about types of interrupts
supported. irq_type callback can be used to validate the types
of interrupt support. If there is additional registers are defined
for level triggered interrupts, end of IRQ callback can be added.

The SMMC driver only implements the generic registers and System
memory as the address space type. Currently, it doesn't support
other address space type but can be easily extended to support it.

Signed-off-by: Atish Patra <atishp@rivosinc.com>